Update
Created foundation for coordinate system and hitboxes for towers. Ceiling function, path now exactly down the middle of the map and other minor changes to the project. Next time I should finish the coordinates logic, print 1 tower on the map using that coordinates system, prescribe some restrictions based on coordinates and tower's hitbox. Make path a separate entity on which explicitly towers can not be placed. Create a numeration of coordinates so that the player is able to see the coordinates in the first place